refactor: break down useAuth hook into modular sub-hooks
Extracted session, login, logout, registration, and email verification
logic into dedicated hooks under hooks/auth to improve readability and
maintainability. The useAuth hook remains a unified orchestrator to
ensure perfect backward compatibility.
